pjt/pjt08/requirements.txt
asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
Django==3.2.8
django-bootstrap-v5==1.0.6
django-extensions==3.1.3
django-seed==0.3.1
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==9.7.1
idna==3.3
inflection==0.5.1
install==1.3.4
ipython==7.28.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
packaging==21.0
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.21
Pygments==2.10.0
pyparsing==3.0.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
